FC-177: The owner of the board can add a column to a board
When a user visits a board they see the columns on the board.  There is a button available to the user to add a new column to the board.

Additional work:
Updated the base ember app kit to include standard dependencies: bootstrap, font-awesome, bootstrap_for_ember.
Initial setup for ember-app-kit

Moved files around to match conventions within Ember.  The card is now a view, and has tests to back that assertion up.
Added the models for cards/columns/boards to project.
Created tests that utilized the new EAK layout.
Note: Fixtures cannot be defined within class definitions.  Cleanup from testing causes the Fixture objects to be marked for destruction.  Thus any data needed or that is tried to be loaded, BESIDES THE ID, is not available.